[REV] decouple console from openapi
